:art: Reorganize files and legacy commands
